feat: moved login context
